Expose ruby version via Gemfile
This pattern has been used on a few repos [1] [2] that have needed
to be run on Heroku, where `.ruby-version` is not respected, and the
version must be set via `Gemfile` instead.

It's not uncommon for us to deploy applications to Heroku for
prototying, and as applications become simpler, and depend on fewer
services (eg, moving to content store) the easier, and more likely
we are to do it in the future.

This commit future-proofs new apps, so they will Just Work in that
situation. It should also act as a canonical record of the pattern,
and our reasons for the choice, rather than both being spread across
many apps.,

We use `ruby-version` over `Gemfile` as it's not dependant on `bundler`,
and also because our target platform (production and VMs) use `rbenv`,
which does not support `.ruby-version`.

[1]: https://github.com/alphagov/govuk-content-schemas/pull/230
[2]: https://github.com/alphagov/government-frontend/blob/master/Gemfile#L3