Remove gem versions from installation script
The installer for a new Rails project probably shouldn't give you a list of outdated gems to get started with.  We encountered some  issues when running the generated application for the first time due to this.

Arguably, the default install should give you the latest versions of your core dependencies.